Convert tiny-process-library in a cmake library
In this way we can use target_link_libraries() to use tiny-process-library
on our own executable targets.

Also we've remove CMAKE_SOURCE_DIR because we need to search for
tiny-process-library source files in the current source dir (instead of
the root global source dir, which could be other if we use
tiny-process-library as a submodule).